current version before last changes

+++ todo.py
@@ -4,7 +4,7 @@ from time import sleep
 # creates a new, empty todo list with headline 'to-do-list'
 def create_new_list():
     with open("list.txt", "w") as file:
-        file.write("To-Do-List")        #todo fix problem with index 
+        file.write("To-Do-List\n")         
 
 # opens the existing list and displays its content
 def open_list():
@@ -29,7 +29,7 @@ def add_item():
                 if var == "y":
                     break
                 elif var == "n":
-                    return
+                    return          #returns to main function
                 else:
                     print("please use 'y' or 'n' to indicate your choice")
                          
@@ -43,18 +43,25 @@ def delete_number():
         lines =file.readlines()
         with open("list.txt", "w") as file:
             for line in lines:
-                parts = line.split(".")
-                clean = parts[-1]
-                file.write(clean)
+                parts = line.split(". ", 1) #splits the lines after the first '. ' to seperate the number from the task. the space behind the'.' is importend
+                if len(parts) > 1 and parts[0].strip().isdigit():   # checks if the line was numbered
+                    clean = parts[1]   #saves only the part after the '.'
+                else:
+                    clean = line # keeps unnumbered lines e.g. headline
+                file.write(clean)   #writes the clean line back
+
 # numbers the tasks based on position in list
 def number_tasks():
     with open("list.txt", "r") as file:
         lines =file.readlines()
+        
         # writes back numbered tasks
         with open("list.txt", "w") as file:
-            for index, line in enumerate(lines, start=1):
-                file.write(f"{index}. {line}")              #todo fix problem with index 
-
+            if lines:
+                file.write(lines[0]) # write first line without number
+                for index, line in enumerate(lines[1:], start=1): # start numbering
+                    file.write(f"{index}. {line}")              
+           
 
 
 # deletes specific lines from list
@@ -64,8 +71,8 @@ def delete_lines(filename, lines_to_delete):
         lines = file.readlines()
     # write back only lines not to be deleted 
     with open("list.txt", "w") as file:
-        for index, line in enumerate(lines, start=1):
-            if index not in lines_to_delete:
+        for index, line in enumerate(lines, start=0):   #iterates through lines
+            if index not in lines_to_delete:            #writes whats not in the lines_to_delete list
                 file.write(line)
 
 def main():
@@ -79,17 +86,21 @@ def main():
     while True:
         print("""1. start a new list\n2. see whats on your to-do-list\n3. add to your to-do-list\n4. delete a completed task from your to-do-list\n5. quit""" )
         try:
-            var = int(input("please press the corresponding number"))
+            var = int(input("please press the corresponding number\n"))
             if var == 1:
                 create_new_list()   #creates a new list
                 add_item()          #lets user add tasks 
                 delete_number()     #cleans the numeration in case user put wrong numbers 
                 number_tasks()      #correctly numbers the tasks
             elif var == 2:
-                open_list()
-                print()
-                print("would you like to add or delete something?")
-                print()
+                try:
+                    open_list()
+                    print()
+                    print("would you like to add or delete something?")
+                    print()
+                except FileNotFoundError:
+                    print("you dont have a to-do-list yet")
+                    print()
             else:
                 if var == 3:
                     add_item()
@@ -102,13 +113,16 @@ def main():
                 elif var == 4:  #todo Handle input errors:
                     open_list()
                     lines_to_delete_input = input("Please enter the numbers of the tasks you'd like to delete, separated by a comma: ") #gets user input as string
-                    lines_to_delete = [int(x.strip()) for x in lines_to_delete_input.split(",") if x.strip().isdigit()] # cuts the string by the','s converts it to int and saves it in a list all in one line very cool!
-                    delete_lines("list.txt", lines_to_delete)
-                    delete_number()
-                    number_tasks()
-                    print("This is your new to-do-list:")
-                    open_list()
-                    print()
+                    if not lines_to_delete_input.isdigit():
+                        print("please select a task using the numbering")
+                    else:
+                        lines_to_delete = [int(x.strip()) for x in lines_to_delete_input.split(",") if x.strip().isdigit()] # cuts the string by the','s converts it to int and saves it in a list all in one line very cool!
+                        delete_lines("list.txt", lines_to_delete)
+                        delete_number()
+                        number_tasks()
+                        print("This is your new to-do-list:")
+                        open_list()
+                        print()
                 else:
                     if var == 5:
                         exit()
